Rework DPS and clean "some" stuff
