+++ src/components/Scrolls.jsx
@@ -1,61 +1,148 @@
 import { useState } from "react";
 import "../style/shop.css";
-export default function Scrolls({
-  handleBuyScroll,
-  handleSellScroll,
-  handleEquipScroll,
-  id,
-  price,
-  bought,
-  equipped,
-  name,
-  handleUse,
-}) {
-  const [isScrollUsed, setIsScrollUsed] = useState(false);
 
-  const handleUseScrolls = () => {
+export default function Scrolls({
+  dealDps,
+  power,
+  score,
+  setScore,
+}) {
+  let [scrolls, setScrolls] = useState([
+    {
+      key: 1,
+      id: 1,
+      bought: false,
+      equipped: false,
+      price: 1000,
+      name: "First Scroll",
+      handleUse: () => {
+        dealDps(power)
+        setTimeout(() => dealDps(-power), 30000);
+      },
+    },
+    {
+      key: 2,
+      id: 2,
+      bought: false,
+      equipped: false,
+      price: 5000,
+      name: "Second Scroll",
+      handleUse: () => {
+        setPower((power = power * 2));
+      },
+    },
+    {
+      key: 3,
+      id: 3,
+      bought: false,
+      equipped: false,
+      price: 8000,
+      name: "Third Scroll",
+    },
+    {
+      key: 4,
+      id: 4,
+      bought: false,
+      equipped: false,
+      price: 15000,
+      name: "Fourth Scroll",
+    },
+  ]);
+  
+  const [isScrollUsed, setIsScrollUsed] = useState(false);
+  const handleUseScrolls = (scroll) => {
     if (isScrollUsed === false) {
-      handleUse();
+      scroll.handleUse();
       setTimeout(setIsScrollUsed(), 3600000);
     }
   };
-  return (
-    <div>
-      {bought === false && (
+
+  const handleBuyScroll = (scroll) => {
+    if (score >= Math.round(scroll.price)) {
+      setScore(score - Math.round(scroll.price));
+      const updatedScrolls = scrolls.map((s) => {
+        if (s.id === scroll.id) {
+          return { ...s, bought: true };
+        }
+        return s;
+      });
+      setScrolls(updatedScrolls);
+    }
+  };
+  const handleSellScroll = (scroll) => {
+    setScore(score + Math.round(scroll.price) / 1.25);
+    const updatedScrolls = scrolls.map((s) => {
+      if (s.id === scroll.id) {
+        return { ...s, equipped: false, bought: false };
+      }
+      return s;
+    });
+    setScrolls(updatedScrolls);
+  };
+  const handleEquipScroll = (scroll, equip) => {
+    let doAction = false;
+    if (!equip) {
+      doAction = true;
+    } else if (equip) {
+      let howManyEquipped = 0;
+      scrolls.forEach((s) => {
+        if (s.equipped) {
+          howManyEquipped++;
+        }
+      });
+      if (howManyEquipped < 3) {
+        doAction = true;
+      }
+    }
+    if (doAction) {
+      const updatedScrolls = scrolls.map((s) => {
+        if (s.id === scroll.id) {
+          return { ...s, equipped: equip };
+        }
+        return s;
+      });
+      setScrolls(updatedScrolls);
+    }
+  };
+
+  return (scrolls.map((s) => {
+    return (
+    <div key={s.key}>
+      {s.bought === false && (
         <button
           className="scrollButtons"
           type="button"
-          onClick={() => handleBuyScroll({ id, price, bought })}
+          onClick={() => handleBuyScroll(s)}
         >
-          Buy {name}: {Math.round(price)} points
+          Buy {s.name}: {Math.round(s.price)} points
         </button>
       )}
-      {bought === true && (
+      {s.bought === true && (
         <div>
           <button
             className="scrollButtons"
             type="button"
-            onClick={() => handleSellScroll({ id, price, bought })}
+            onClick={() => handleSellScroll(s)}
           >
-            Sell {name}: You will gain {Math.round(price) / 1.25} points
+            Sell {s.name}: You will gain {Math.round(s.price) / 1.25} points
           </button>
-          {equipped === false && (
+          {s.equipped === false && (
             <button
               type="button"
-              onClick={() => handleEquipScroll({ id, price, bought }, true)}
+              onClick={() => handleEquipScroll(s, true)}
             >
               Equip
             </button>
           )}
-          {equipped === true && (
+          {s.equipped === true && (
             <div>
               <button
                 type="button"
-                onClick={() => handleEquipScroll({ id, price, bought }, false)}
+                onClick={() => handleEquipScroll(s, false)}
               >
                 Unequip
               </button>
-              <button type="button" onClick={() => handleUseScrolls()}>
+              <button type="button" onClick={() => handleUseScrolls(s)}>
                 Use
               </button>
             </div>
@@ -63,5 +150,6 @@ export default function Scrolls({
         </div>
       )}
     </div>
-  );
+    )
+  }));
 }
